refactor :: 컨플릭트 수정

+++ app/src/main/java/team/retum/jobisandroidv2/root/RootNavigation.kt
@@ -8,7 +8,7 @@ const val NAVIGATION_ROOT = "root"
 
 fun NavGraphBuilder.root(
     onAlarmClick: () -> Unit,
-    onRecruitmentDetailsClick: () -> Unit,
+    onRecruitmentDetailsClick: (Long) -> Unit,
 ) {
     composable(NAVIGATION_ROOT) {
         Root(
